requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
packaging==23.2
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.18.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
